requirements.txt
pandas>=1.5.0
numpy>=1.21.0
google-adk>=0.1.0
google-genai>=0.3.0
asyncio
typing-extensions>=4.0.0